Get tests running on wasm32-wasi
We need to use a fork of splitmix here. We /should/ be able to still
point to the original repo, but Cabal doesn't run a "git fetch" to
download the ref, so you can't point to a commit not on a branch.
This is fixed in latest, but that won't be released for a while.