Skip major/minor/makedev on WASM (#16)
* Skip major/minor/makedev on WASM

* Always bundle include files

We just skip compiling it. Yes, this is horrible, but what can you do.

* Remove NEED_setSymbolicLinkOwnerAndGroup

Alternative solution to building. This allows us to remove the include
of HsUnixCompat.

* Get tests running on wasm32-wasi

We need to use a fork of splitmix here. We /should/ be able to still
point to the original repo, but Cabal doesn't run a "git fetch" to
download the ref, so you can't point to a commit not on a branch.
This is fixed in latest, but that won't be released for a while.

* Build/test against wasm32-wasi

GHCUp doesn't appear to have GHC 9.12. While we don't strictly speaking
need it here (it would be sufficient to test with GHC 9.10), it feels
easier to just rely on the Nix flake.

* Add changelog entry